ArangoDB: fixed driver lifecycle
