+++ jnosql-arangodb/src/main/java/org/eclipse/jnosql/databases/arangodb/communication/ArangoDBBucketManager.java
@@ -132,7 +132,7 @@ public class ArangoDBBucketManager implements BucketManager, ArangoDBAccessor {
 
     @Override
     public void close() {
-
+        arangoDB.shutdown();
     }
 
     @Override
